Separate vLLM and Transformers MoE gates
